fixes #6: Using XMPP address as email address should be configurable
The plugin will try to look up the email address of senders and recipients of chat messages, but an email address can not always be found. The system property "plugin.emailonaway.use_xmpp_as_email_address" defines if the (bare) XMPP address of a user is to be used as an email address if no other address can be found. When configured to 'false', then the value from the system property "plugin.emailonaway.email.default_mail_address" will be used.

To implement the above, a refactoring has been implemented for the lookup of both the 'human name' as well as email address for an XMPP address.